feat: add new parent's option to clone (#11)
* feat: add new parent's option to clone

Decide whether to clone the parent or not as a base for the child
auto-created channel. Defaults to false.

* edit: remove option parameters when using childShouldBeACopyOfParent: true (#12)

This would result that the channel will be a full clone of the parent channel exept the naming pattern. In my opinion  this might give that new added parameter more use of it.

* docs: update API doc for new parameter

* chore: add build workflow

---------

Co-authored-by: HunteRoi <me@tinaeldevresse.eu>
Co-authored-by: Alexander Ballauf <38478976+Ballaual@users.noreply.github.com>